1.7.5-DEVELOPMENT-SNAPSHOT

+++ src/main/java/com/gmail/gogobebe2/shiftkits/MagicKit.java
@@ -10,12 +10,12 @@ import java.util.Map;
 
 public class MagicKit extends Kit implements Listener {
 
-    public MagicKit(String name, int level, Requirement requirement, Map<Integer, ItemStack> contents, Material helmet, Material chestplate, Material leggings, Material boots, Material icon, Listener listener) {
+    public MagicKit(String name, short level, Requirement requirement, Map<Integer, ItemStack> contents, Material helmet, Material chestplate, Material leggings, Material boots, Material icon, Listener listener) {
         super(name, level, requirement, contents, helmet, chestplate, leggings, boots, icon);
         Bukkit.getPluginManager().registerEvents(listener, ShiftKits.instance);
     }
 
-    public MagicKit(String name, int level, Requirement requirement, Map<Integer, ItemStack> contents, Material icon, Listener listener) {
+    public MagicKit(String name, short level, Requirement requirement, Map<Integer, ItemStack> contents, Material icon, Listener listener) {
         this(name, level, requirement, contents, null, null, null, null, icon, listener);
     }
 }
